Moved common user test functions to util
